Convidados.vt5

+++ app/src/main/java/com/example/convidados/view/viewholder/GuestViewHolder.kt
@@ -1,6 +1,8 @@
 package com.example.convidados.view.viewholder
 
+import android.content.DialogInterface
 import android.view.View
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.example.convidados.databinding.RowGuestBinding
 import com.example.convidados.model.GuestModel
@@ -16,8 +18,18 @@ class GuestViewHolder(private val bind: RowGuestBinding, private val listener: O
             listener.onClick(guest.id)
         }
 
+        // Implementação da caixinha de confirmação
+
         bind.textName.setOnLongClickListener {
-            listener.onDelete(guest.id)
+            AlertDialog.Builder(itemView.context)
+                .setTitle("Remoção de convidado")
+                .setMessage("Tem certeza que deseja remover?")
+                .setPositiveButton("Sim") { dialog, which ->
+                    listener.onDelete(guest.id)
+                }
+                .setNegativeButton("Não", null)
+                .create()
+                .show()
             true
         }
